requirements.txt
Flask==3.1.0
gunicorn==23.0.0
flask-cors==5.0.1
PyMySQL==1.1.1
python-jobspy==1.1.78
aiohttp==3.11.14
langdetect==1.0.9
tqdm==4.67.1
flask_apscheduler==1.13.1
nltk==3.9.1
scikit-learn==1.6.1
